move code to service layer from routes
